Refactor performance optimizer into modular ES6 module
