+++ public/index.html
@@ -532,6 +532,6 @@
       }
     });
   </script>
-  <script src="./main-simple.js?v=23&debug=true"></script>
+  <script type="module" src="./main-simple.js?v=23&debug=true"></script>
 </body>
 </html>
